Use `round` instead of `trunc`.
The function `trunc` in NESL seems to be doing something different
than `truncate` in Haskell. If we use `trunc` then `sum(dists)` is
always 0, now we at least converges towards π. Alas, we are not
converging as fast as the Haskell version, which worries me because I
though that the algorithm was deterministic.